# Backend/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "University Evaluation API",
    "version": "1.0.0",
    "description": "University evaluation system with 3 dimensions: Governance, Social, Environmental - Auto-generated documentation",
    "contact": {
      "name": "API Support",
      "email": "support@api.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "LOGIN_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Invalid credentials"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/dimensions": {
      "get": {
        "tags": [
          "Dimensions"
        ],
        "summary": "Get all evaluation dimensions",
        "description": "Retrieve the list of all evaluation dimensions (Governance, Social, Environmental)",
        "responses": {
          "200": {
            "description": "List of dimensions retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Dimensions retrieved successfully"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer",
                            "example": 1
                          },
                          "name": {
                            "type": "string",
                            "example": "Governance"
                          },
                          "code": {
                            "type": "string",
                            "example": "governance"
                          },
                          "description": {
                            "type": "string",
                            "example": "Governance dimension evaluation"
                          },
                          "created_at": {
                            "type": "string",
                            "format": "date-time",
                            "example": "2024-01-01T10:00:00Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/dimensions/{id}": {
      "get": {
        "tags": [
          "Dimensions"
        ],
        "summary": "Get a specific dimension",
        "description": "Retrieve details of a specific evaluation dimension by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID of the dimension to retrieve",
            "example": 1
          }
        ],
        "responses": {
          "200": {
            "description": "Dimension retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Dimension retrieved successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer",
                          "example": 1
                        },
                        "name": {
                          "type": "string",
                          "example": "Governance"
                        },
                        "code": {
                          "type": "string",
                          "example": "governance"
                        },
                        "description": {
                          "type": "string",
                          "example": "Governance dimension evaluation"
                        },
                        "created_at": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2024-01-01T10:00:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Dimension not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}